Rename function to generatePncExceptionFromMessage
